refactor: move file dialogs up to composite
